typing everything without utils
